decomp: Singer::HandlePhraseEnd 92.7->93.4%
Move mResultsData[i].unk14=0 below unk10=0 inside the loop and
mAmbiguousData.clear() to the function end, matching the target's store
ordering. Behavior-neutral (neither field is read in the loop).
Remainder is r0<->r6 coloring in the inlined stlport clear/erase loop
(at-limit).

+++ src/system/rndobj/MeshDeform.cpp
@@ -299,7 +299,7 @@ void RndMeshDeform::Reskin(SyncMeshCB *cb, bool force) {
             Multiply(mBones[i].unk14, tmp, xfms[i]);
         } else {
             xfms[i].Reset();
-            TheDebug << MakeString("%s null bone %d\n", PathName(this), i);
+            TheDebug << MakeString("%s null bone %d\n", PathName(this), (int)i);
         }
     }
     int meshNumVerts = mMesh->Verts().size();
